backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
pydantic==2.5.0
sqlalchemy==2.0.23
# alembic==1.12.1  # Database migrations (for future use)

# AWS SDK
boto3==1.29.0
botocore==1.32.0

# Database drivers
# asyncpg==0.29.0  # Async PostgreSQL driver (for future PostgreSQL migration)
# psycopg2-binary==2.9.9  # PostgreSQL (sync fallback)
# pymysql==1.1.0  # MySQL (not currently used)

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
bleach==6.1.0  # HTML sanitization

# HTTP client
httpx==0.25.0
requests==2.31.0

# Task Queue & Caching
celery==5.3.4
redis==5.0.1
# flower==2.0.1  # Celery monitoring (optional - run manually if needed)

# Rate Limiting & Circuit Breaker
slowapi==0.1.9  # Rate limiting for FastAPI
pybreaker==1.0.1  # Circuit breaker pattern
tenacity==8.2.3  # Retry with exponential backoff

# CSV Processing
pandas==2.1.4
aiofiles==23.2.1  # Async file operations

# AI Services (install as needed based on which service you want to use)
# Google Gemini
google-generativeai==0.3.2

# OpenAI (optional)
# openai==1.6.1

# Anthropic Claude (optional)
# anthropic==0.8.1

# Cohere (optional)
# cohere==4.37

# HuggingFace (optional)
# huggingface-hub==0.20.1

# Groq (optional)
# groq==0.4.1
